Tachzukanit ========= - getting the long and lat of every apartment by its address

+++ Tachzukanit/TachzukanitBE/Models/Apartment.cs
@@ -18,9 +18,9 @@ namespace TachzukanitBE.Models
         [MaxLength(50)]
         public string Address { get; set; }
 
-        public float Longitude { get; set; }
+        public double Longitude { get; set; }
 
-        public float Latitude { get; set; }
+        public double Latitude { get; set; }
 
         public string Photo { get; set; }
